Split Pulse diagnostics hook
